Migrate all searches page code to page component in features/search
