+++ features/search/components/twitch-search/index.ts
@@ -1,2 +1,3 @@
+export { TwitchSearchPage } from "./TwitchSearchPage";
 export { TwitchSearchTab } from "./TwitchSearchTab";
 export { TwitchChannelResult } from "./TwitchChannelResult";
