+++ features/search/components/youtube-search/index.ts
@@ -1,2 +1,3 @@
+export { YouTubeSearchPage } from "./YouTubeSearchPage";
 export { YouTubeChannelResult } from "./YouTubeChannelResult";
 export { YouTubeSearchTab } from "./YouTubeSearchTab";
